+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- A base class for Custom Mappers and IRecipe mappers.
+- Shortcut for NSSOutput
+
 ### Changed
 - Made getDesc an optional override for ARecipeTypeMapper.
-- Renamed [Extended Crafting] mappers 
+- Renamed [Extended Crafting] mappers
+- Moved RecipeTypeMapper getDesc to base class
 
 ### Removed
 - All overridden getDesc in all addons.